cloud_run_model_service/app/requirements.txt
Flask
gunicorn
tensorflow
numpy
google-cloud-storage